+++ apps/patient/src/views/ReadyBy.tsx
@@ -5,12 +5,12 @@ import {
   Card,
   CardBody,
   Container,
+  Fade,
   Heading,
   HStack,
   Icon,
   Radio,
   RadioGroup,
-  Fade,
   Tag,
   Text,
   VStack
@@ -20,14 +20,14 @@ import { Helmet } from 'react-helmet';
 import dayjs from 'dayjs';
 import { datadogRum } from '@datadog/browser-rum';
 import timezone from 'dayjs/plugin/timezone';
-import { capitalize, convertReadyByToUTCTimestamp } from '../utils/general';
-
-dayjs.extend(timezone);
-
+import { convertReadyByToUTCTimestamp } from '../utils/general';
 import { FixedFooter, PoweredBy } from '../components';
 import { text as t } from '../utils/text';
 import { useOrderContext } from './Main';
 import { RxLightningBolt } from 'react-icons/rx';
+import { capitalize } from '../utils/formatters';
+
+dayjs.extend(timezone);
 
 const checkDisabled = (option: string): boolean => {
   const currentTime = dayjs();
