+++ apps/patient/src/views/Status.tsx
@@ -13,7 +13,6 @@ import { OrderSummary } from '../components/order-summary/OrderSummary';
 import { OrderStatusHeader } from '../components/status/Header';
 import { deriveOrderStatus, getLatestReadyTime } from '../utils/fulfillmentsHelpers';
 import {
-  formatAddress,
   getFulfillmentType,
   isDelivery,
   isRerouteablePharmacy,
@@ -22,6 +21,7 @@ import {
 import { InsuranceAlert } from '../components/InsuranceAlert';
 import { text as t } from '../utils/text';
 import { useOrderContext } from './Main';
+import { formatAddress } from '../utils/formatters';
 
 export const Status = () => {
   const navigate = useNavigate();
